// src/app/modules/favorites/store/types/favorites.types.ts
export enum Types {
    CREATE = '[Favorites] createCategory',
    CREATE_ERROR = '[Favorites] createCategoryError',
    CREATE_SUCCESS = '[Favorites] createCategorySuccess',
    DELETE = '[Favorites] deleteCategory',
    DELETE_ERROR = '[Favorites] deleteCategoryError',
    DELETE_SUCCESS = '[Favorites] deleteCategorySuccess',
    LOAD = '[Favorites] loadFavorites',
    LOAD_ERROR = '[Favorites] loadFavoritesError',
    LOAD_SUCCESS = '[Favorites] loadFavoritesSuccess',
    UPDATE = '[Favorites] updateCategory',
    UPDATE_ERROR = '[Favorites] updateCategoryError',
    UPDATE_SUCCESS = '[Favorites] updateCategorySuccess'
}